day08ver1.0

+++ pinyougou-parent/pinyougou-sellergoods-service/src/main/java/com/pinyougou/sellergoods/service/impl/ItemCatServiceImpl.java
@@ -1,6 +1,11 @@
 package com.pinyougou.sellergoods.service.impl;
+
 import java.util.List;
+
+import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
+
 import com.alibaba.dubbo.config.annotation.Service;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
@@ -14,15 +19,17 @@ import entity.PageResult;
 
 /**
  * 服务实现层
+ * 
  * @author Administrator
  *
  */
 @Service
+@Transactional
 public class ItemCatServiceImpl implements ItemCatService {
 
 	@Autowired
 	private TbItemCatMapper itemCatMapper;
-	
+
 	/**
 	 * 查询全部
 	 */
@@ -36,8 +43,8 @@ public class ItemCatServiceImpl implements ItemCatService {
 	 */
 	@Override
 	public PageResult findPage(int pageNum, int pageSize) {
-		PageHelper.startPage(pageNum, pageSize);		
-		Page<TbItemCat> page=   (Page<TbItemCat>) itemCatMapper.selectByExample(null);
+		PageHelper.startPage(pageNum, pageSize);
+		Page<TbItemCat> page = (Page<TbItemCat>) itemCatMapper.selectByExample(null);
 		return new PageResult(page.getTotal(), page.getResult());
 	}
 
@@ -46,25 +53,25 @@ public class ItemCatServiceImpl implements ItemCatService {
 	 */
 	@Override
 	public void add(TbItemCat itemCat) {
-		itemCatMapper.insert(itemCat);		
+		itemCatMapper.insert(itemCat);
 	}
 
-	
 	/**
 	 * 修改
 	 */
 	@Override
-	public void update(TbItemCat itemCat){
+	public void update(TbItemCat itemCat) {
 		itemCatMapper.updateByPrimaryKey(itemCat);
-	}	
-	
+	}
+
 	/**
 	 * 根据ID获取实体
+	 * 
 	 * @param id
 	 * @return
 	 */
 	@Override
-	public TbItemCat findOne(Long id){
+	public TbItemCat findOne(Long id) {
 		return itemCatMapper.selectByPrimaryKey(id);
 	}
 
@@ -72,29 +79,51 @@ public class ItemCatServiceImpl implements ItemCatService {
 	 * 批量删除
 	 */
 	@Override
-	public void delete(Long[] ids) {
-		for(Long id:ids){
-			itemCatMapper.deleteByPrimaryKey(id);
-		}		
+	public long delete(Long[] ids) {
+		System.out.println("ipml进来了");
+		for (Long id : ids) {
+			TbItemCatExample example = new TbItemCatExample();
+			Criteria criteria = example.createCriteria();
+			criteria.andParentIdEqualTo(id);
+			List<TbItemCat> list = itemCatMapper.selectByExample(example);
+			System.out.println(list);
+			if(list.toString() != "[]") {
+				return id;
+			}else {
+				itemCatMapper.deleteByPrimaryKey(id);
+				System.out.println(id);
+			}
+		}
+		return 0;
 	}
-	
-	
-		@Override
+
+	@Override
 	public PageResult findPage(TbItemCat itemCat, int pageNum, int pageSize) {
 		PageHelper.startPage(pageNum, pageSize);
-		
-		TbItemCatExample example=new TbItemCatExample();
+
+		TbItemCatExample example = new TbItemCatExample();
 		Criteria criteria = example.createCriteria();
-		
-		if(itemCat!=null){			
-						if(itemCat.getName()!=null && itemCat.getName().length()>0){
-				criteria.andNameLike("%"+itemCat.getName()+"%");
+
+		if (itemCat != null) {
+			if (itemCat.getName() != null && itemCat.getName().length() > 0) {
+				criteria.andNameLike("%" + itemCat.getName() + "%");
 			}
-	
+
 		}
-		
-		Page<TbItemCat> page= (Page<TbItemCat>)itemCatMapper.selectByExample(example);		
+
+		Page<TbItemCat> page = (Page<TbItemCat>) itemCatMapper.selectByExample(example);
 		return new PageResult(page.getTotal(), page.getResult());
 	}
 	
+	/**
+	 * 查找下级目录
+	 */
+	@Override
+	public List<TbItemCat> findByParentId(Long parentId) {
+		TbItemCatExample example = new TbItemCatExample();
+		Criteria criteria = example.createCriteria();
+		criteria.andParentIdEqualTo(parentId);
+		return itemCatMapper.selectByExample(example);
+	}
+
 }
